modules/video_assembler.py
import logging
import ffmpeg
from pathlib import Path
from typing import Optional, Dict, Any

logger = logging.getLogger(__name__)

class VideoAssembler:
    """
    Assembles the final 4K video using FFmpeg, layering vocals, ambient drones,
    and looping background visuals.
    """
    def __init__(self, output_dir: str = "output/rendered_videos"):
        self.output_dir = Path(output_dir)
        self.output_dir.mkdir(parents=True, exist_ok=True)

    def render_final_video(
        self, 
        audio_path: str, 
        background_path: str, 
        duration_hours: int,
        output_filename: str = "final_render.mp4"
    ) -> str:
        """
        Renders the final video with looped background and layered audio.
        """
        output_path = self.output_dir / output_filename
        duration_seconds = duration_hours * 3600
        
        logger.info(f"Starting final render: {output_path} ({duration_hours} hours)")
        
        try:
            # Input background loop
            video_input = ffmpeg.input(background_path, stream_loop=-1, t=duration_seconds)
            
            # Input processed audio
            audio_input = ffmpeg.input(audio_path)
            
            # Assembly
            stream = ffmpeg.output(
                video_input,
                audio_input,
                str(output_path),
                vcodec='libvpx-vp9',
                video_bitrate='20M',
                acodec='libopus',
                audio_bitrate='128k',
                crf=30,
                s='3840x2160',
                r=30,
                pix_fmt='yuv420p',
                **{'b:v': '20M', 'maxrate': '25M', 'bufsize': '50M'}
            ).global_args('-threads', '8')
            
            # In a real scenario, we would use .run_async() to track progress
            # For this implementation, we'll use .run()
            logger.info("Executing FFmpeg command...")
            stream.run(overwrite_output=True)
            
            logger.info(f"Render complete: {output_path}")
            return str(output_path)
            
        except ffmpeg.Error as e:
            logger.error(f"FFmpeg error: {e.stderr.decode()}")
            raise

    def add_text_overlay(self, video_path: str, text: str, start_time: float, duration: float) -> str:
        """
        Adds a text overlay to a video segment.
        (Simplified implementation for the pipeline)
        """
        logger.info(f"Adding text overlay: '{text}' at {start_time}s")
        # Implementation would involve ffmpeg.drawtext filter
        return video_path
